Break Out DataCoordinator into separate file
