+++ custom_components/teamtracker/coordinator.py
@@ -1,0 +1,581 @@
+""" TeamTracker Data Coordinator """
+from datetime import date, datetime, timedelta, timezone
+import locale
+import logging
+import re
+
+import arrow
+from async_timeout import timeout
+from typing import ClassVar
+
+from homeassistant.config_entries import ConfigEntry
+from homeassistant.const import CONF_NAME
+from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
+
+from .clear_values import async_clear_values
+
+from .const import (
+    API_LIMIT,
+    CONF_API_LANGUAGE,
+    CONF_CONFERENCE_ID,
+    CONF_LEAGUE_ID,
+    CONF_LEAGUE_PATH,
+    CONF_SPORT_PATH,
+    CONF_TEAM_ID,
+    DEFAULT_LOGO,
+    DEFAULT_TIMEOUT,
+    URL_HEAD,
+    URL_TAIL,
+)
+from .event import async_process_event
+from .hockeytech import (
+    async_fetch_hockeytech_data,
+)
+from .provider_factory import (
+    DATA_PROVIDER_ESPN, 
+    DATA_PROVIDER_ESPN_ALL_LEAGUES, 
+    DATA_PROVIDER_HOCKEYTECH, 
+    get_provider)
+from .utils import is_integer, async_call_espn_api, has_team
+
+_LOGGER = logging.getLogger(__name__)
+
+def _slug_to_name(slug: str) -> str:
+    """Convert a season slug like '2025-26-english-premier-league' to 'English Premier League'."""
+    if not slug:
+        return ""
+    body = re.sub(r"^\d{4}(-\d{2})?-", "", slug)
+    if body == slug:
+        return ""
+    def _fmt_word(w):
+        # Uppercase abbreviations (no vowels, e.g. "mls", "nfl"); title-case real words
+        return w.upper() if w.isalpha() and not re.search(r"[aeiou]", w, re.I) else w.title()
+    return " ".join(_fmt_word(w) for w in body.split("-"))
+
+
+class TeamTrackerCoordinator(DataUpdateCoordinator):
+    """Class to manage fetching TeamTracker data."""
+
+# Stores API data for sharing across sensors
+#  key = "{sport_path}:{league_path}:{conference_id}:{lang}"+":{team_id}" if league_path "all"
+    data_cache: ClassVar[dict] = {}  # {key: {cache_data, cache_url, cache_time}}
+
+# Stores team information when league_path is all
+#  key = "{sport}:{league}:{team_id}"
+    all_team_cache: ClassVar[dict] = {}  # {key: {next_game_date, league_map, expires}}
+
+    def __init__(self, hass, config, entry: ConfigEntry=None):
+        """Initialize."""
+        self.api_url = ""
+
+        self.name = config[CONF_NAME]
+        self.team_id = config[CONF_TEAM_ID]
+        self.league_id = config[CONF_LEAGUE_ID]
+        self.league_path = config[CONF_LEAGUE_PATH]
+        self.sport_path = config[CONF_SPORT_PATH]
+        self.conference_id = ""
+        if CONF_CONFERENCE_ID in config.keys():
+            if len(config[CONF_CONFERENCE_ID]) > 0:
+                self.conference_id = config[CONF_CONFERENCE_ID]
+
+        if self.sport_path.lower() == DATA_PROVIDER_HOCKEYTECH:
+            self.provider = get_provider(DATA_PROVIDER_HOCKEYTECH)
+        elif self.league_path.lower() == "all":
+            self.provider = get_provider(DATA_PROVIDER_ESPN_ALL_LEAGUES)
+        else:
+            self.provider = get_provider(DATA_PROVIDER_ESPN)
+
+        self.update_interval = self.provider.DEFAULT_REFRESH_RATE
+
+        self.config = config
+        self.hass = hass
+        self.entry = entry #None if setup from YAML
+
+        super().__init__(hass, _LOGGER, name=self.name, update_interval=self.provider.DEFAULT_REFRESH_RATE)
+        _LOGGER.debug(
+            "%s: Using default refresh rate (%s)", self.name, self.update_interval
+        )
+
+
+    #
+    #  Return the language to use for the API
+    #
+    def get_lang(self):
+        """Return language to use for API."""
+
+        try:
+            lang = self.hass.config.language
+        except:
+            lang, _ = locale.getlocale()
+            lang = lang or "en_US"
+
+        # Override language if is set in the configuration or options
+
+        if CONF_API_LANGUAGE in self.config.keys():
+            lang = self.config[CONF_API_LANGUAGE].lower()
+        if self.entry and self.entry.options and CONF_API_LANGUAGE in self.entry.options and len(self.entry.options[CONF_API_LANGUAGE])>=2:
+                lang = self.entry.options[CONF_API_LANGUAGE].lower()
+
+        return lang
+
+
+    #
+    #  Set team info from service call
+    #
+    def update_team_info(self, sport_path, league_path, team_id, conference_id=""):
+        """update team information when call_api service is called."""
+
+        self.sport_path = sport_path
+        self.league_path = league_path
+        self.league_id = "XXX"
+        self.team_id = team_id
+        self.conference_id = conference_id
+
+        lang = self.get_lang()
+        key = sport_path + ":" + league_path + ":" + conference_id + ":" + lang
+        if league_path == "all" and is_integer(self.team_id):
+            key += ":" + team_id
+
+        if key in TeamTrackerCoordinator.data_cache:
+            TeamTrackerCoordinator.data_cache.pop(key, None)
+
+
+    #
+    #  async_get_team_schedule()
+    #
+    #    Calls the team info and schedule endpoints to discover the next game
+    #    date and build an event_id → league name mapping (substring of season)
+    #    Results are cached in all_team_cache until the next game date passes.
+    #
+    async def async_get_team_schedule(self):
+        """Fetch team schedule info for 'all' league date computation."""
+
+        team_id = self.team_id
+        sport_path = self.sport_path
+        league_path = self.league_path
+        sensor_name = self.name
+
+        cache_key = f"{sport_path}:{league_path}:{team_id}"
+        today = date.today()
+        cached = TeamTrackerCoordinator.all_team_cache.get(cache_key)
+
+        if cached is not None and today <= cached["expires"]:
+            _LOGGER.debug("%s: all_team_cache hit for '%s'", sensor_name, team_id)
+            return cached
+
+        team_url = URL_HEAD + sport_path + "/" + league_path + "/teams/" + team_id
+
+        league_map = {}
+        next_events = []
+
+        response = await async_call_espn_api(self.hass, team_url, None, sensor_name, team_id)
+        team_data = response["data"]
+        if team_data:
+            next_events = team_data.get("team", {}).get("nextEvent", [])
+            for ne in next_events:
+                eid = ne.get("id")
+                if not eid:
+                    continue
+                display = ne.get("season", {}).get("displayName") or _slug_to_name(
+                    ne.get("season", {}).get("slug", "")
+                )
+                if display:
+                    league_map[str(eid)] = display
+
+        schedule_url = team_url + "/schedule"
+        response = await async_call_espn_api(self.hass, schedule_url, None, sensor_name, team_id)
+        sched_data = response["data"]
+        if sched_data:
+            for e in sched_data.get("events", []):
+                eid = e.get("id")
+                if not eid:
+                    continue
+                display = e.get("season", {}).get("displayName") or _slug_to_name(
+                    e.get("season", {}).get("slug", "")
+                )
+                if display:
+                    league_map[str(eid)] = display
+
+        next_game_date = (
+            date.fromisoformat(next_events[0]["date"][:10]) if next_events else None
+        )
+
+        result = {
+            "next_game_date": next_game_date,
+            "league_map": league_map,
+            "expires": next_game_date or today,
+        }
+        TeamTrackerCoordinator.all_team_cache[cache_key] = result
+        return result
+
+
+    #
+    #  DataUpdateCoordinator Call Tree
+    #
+    #  _async_update_data() - Top-level method called from HA to update sensor, controls refresh rate
+    #    async_update_sport_data() - Determines to use cached data or API call (if exprired)
+    #      async_call_sport_apis() - Calls appropriate set of APIs based on sport and league
+    #        async_fetch_espn_data() - Gets data from ESPN APIs for specified league
+    #          async_call_espn_api() - Mockable, overridable API call for ESPN APIs
+    #        async_fetch_espn_all_leagues_data() - Gets data from ESPN APIs for all leagues in specified sport
+    #          async_call_espn_api() - Mockable, overridable API call for ESPN APIs
+    #        async_fetch_hockeytech_data() - Gets data from HockeyTech APIs for specified league
+    #          async_call_hockeytech_api() - Mockable, overridable API call for HockeyTech APIs
+    #      async_update_values() - Updates sensor values using data returned by API or in cache
+    #        async_process_event() - Parses ESPN event structure and populates values for sensor
+    #
+    async def _async_update_data(self):
+        """Top-level method called from HA to update sensor, controls refresh rate."""
+        async with timeout(DEFAULT_TIMEOUT):
+            try:
+                data = await self.async_update_sport_data()
+
+                # update the interval based on flag
+                if data["private_fast_refresh"]:
+                    if self.update_interval != self.provider.RAPID_REFRESH_RATE:
+                        self.update_interval = self.provider.RAPID_REFRESH_RATE
+                        _LOGGER.debug(
+                            "%s: Switching to rapid refresh rate (%s)", self.name, self.update_interval
+                        )
+                else:
+                    if self.update_interval != self.provider.DEFAULT_REFRESH_RATE:
+                        self.update_interval = self.provider.DEFAULT_REFRESH_RATE
+                        _LOGGER.debug(
+                            "%s: Switching to default refresh rate (%s)", self.name, self.update_interval
+                        )
+            except Exception as error:
+                _LOGGER.debug("%s: Error updating data: %s", self.name, error)
+                _LOGGER.debug("%s: Error type: %s", self.name, type(error).__name__)
+                _LOGGER.debug("%s: Additional information: %s", self.name, str(error))
+                raise UpdateFailed(error) from error
+            return data
+
+#
+#  async_update_sport_data()
+#
+    async def async_update_sport_data(self) -> dict:
+        """Determines to use cached data or API call (if exprired)"""
+
+        sport_path = self.sport_path
+        league_path = self.league_path
+        conference_id = self.conference_id
+
+        lang = self.get_lang()
+
+        # For "all" leagues, include team_id in cache key since each team
+        # uses different narrow date windows for the scoreboard call.
+        key = sport_path + ":" + league_path + ":" + conference_id + ":" + lang
+        if league_path == "all" and is_integer(self.team_id):
+            key += ":" + self.team_id
+
+        #
+        #  Use cache if not expired
+        #
+        dc = TeamTrackerCoordinator.data_cache.get(key, None)
+        if dc:
+            cache_time = dc.get("cache_time", None)
+
+            expiration = (
+                datetime.fromisoformat(cache_time) + self.update_interval
+            )
+            now = datetime.now(timezone.utc)
+
+            if now < expiration:
+                data = dc.get("cache_data", None)
+                self.api_url = dc.get("cache_url", None)
+
+                values = await self.async_update_values(data)
+
+                if values["api_message"]:
+                    values["api_message"] = "Cached data: " + values["api_message"]
+                else:
+                    values["api_message"] = "Cached data"
+
+                return values
+
+        data = await self.async_call_sport_apis()
+        values = await self.async_update_values(data)
+
+        if data is not None:
+            TeamTrackerCoordinator.data_cache[key] = {
+                "cache_data": data,
+                "cache_url": self.api_url,
+                "cache_time": values["last_update"]
+            }
+
+        return values
+
+
+    #
+    #  async_call_sport_apis()
+    #    This is the API dispatcher, calls to new non-ESPN API's should be added here based on league_path.
+    #      Response data should be formatted as an ESPN event.
+    #
+    async def async_call_sport_apis(self) -> dict:
+        """Calls appropriate set of APIs based on sport and league."""
+
+        lang = self.get_lang()
+        league_path = self.league_path
+        if self.provider.DATA_PROVIDER == DATA_PROVIDER_HOCKEYTECH:
+            response = await async_fetch_hockeytech_data(self.hass, league_path.upper(), self.name, lang)
+        elif (league_path == "all") and is_integer(self.team_id):
+            response = await self.async_fetch_espn_all_leagues_data(self.hass, lang)
+        else:
+            response = await self.async_fetch_espn_data(self.hass, lang)
+
+        self.api_url = response["url"]
+        return response["data"]
+
+
+    #
+    #  async_fetch_espn_data()
+    #    Call ESPN API with using varying date ranges and parameters until events returned
+    #      1. Call w/ sport specific date range
+    #      2. Call w/o date range specfied (uses ESPN default behavior)
+    #      3. Call w/o language parm (some sports not returned in some languages)
+    #
+    async def async_fetch_espn_data(self, hass, lang) -> dict:
+        """Gets data from ESPN APIs for specified league."""
+
+        sensor_name = self.name
+        sport_path = self.sport_path
+        league_path = self.league_path
+        team_id = self.team_id.upper()
+
+        url_parms = {}
+        url_parms["lang"] = lang[:2]
+        url_parms["limit"] = str(API_LIMIT)
+
+        if sport_path not in ("tennis"):
+            d1 = (date.today() - timedelta(days=1)).strftime("%Y%m%d")
+            if league_path == "all":
+                d2 = (date.today() + timedelta(days=5)).strftime("%Y%m%d")
+            elif sport_path in ("baseball"):
+                d2 = (date.today() + timedelta(days=1)).strftime("%Y%m%d")
+            else:
+                d2 = (date.today() + timedelta(days=90)).strftime("%Y%m%d")
+            url_parms["dates"] = f"{d1}-{d2}"
+
+        file_override = False
+        if self.conference_id:
+            url_parms["groups"] = self.conference_id
+            if self.conference_id == "9999":
+                file_override = True
+
+        url = URL_HEAD + sport_path + "/" + league_path + URL_TAIL
+
+        response = await async_call_espn_api(hass, url, url_parms, sensor_name, team_id, file_override)
+        data = response["data"]
+        url = response["url"]
+
+        num_events = 0
+        if data is not None:
+            _LOGGER.debug(
+                "%s: Data returned for '%s' from %s",
+                sensor_name,
+                team_id,
+                url,
+            )
+            try:
+                num_events = len(data["events"])
+            except:
+                num_events = 0
+
+        _LOGGER.debug(
+            "%s: Num_events '%d' from %s",
+            sensor_name,
+            num_events,
+            url,
+        )
+            
+        # First fallback - without date constraint
+        if num_events == 0:
+            url_parms.pop("dates", None)
+
+            url = URL_HEAD + sport_path + "/" + league_path + URL_TAIL
+
+            response = await async_call_espn_api(hass, url, url_parms, sensor_name, team_id)
+            data = response["data"]
+            url = response["url"]
+
+            num_events = 0
+            if data is not None:
+                _LOGGER.debug(
+                    "%s: Data returned for '%s' from %s",
+                    sensor_name,
+                    team_id,
+                    url,
+                )
+                try:
+                    num_events = len(data["events"])
+                except:
+                    num_events = 0
+
+            _LOGGER.debug(
+                "%s: Num_events '%d' from %s",
+                sensor_name,
+                num_events,
+                url,
+            )
+
+        # Second fallback - without language
+        if num_events == 0:
+            url_parms.pop("lang", None)
+
+            url = URL_HEAD + sport_path + "/" + league_path + URL_TAIL
+            _LOGGER.debug(
+                "%s: Calling API without language for '%s' from %s",
+                sensor_name,
+                team_id,
+                url,
+            )
+
+            response = await async_call_espn_api(hass, url, url_parms, sensor_name, team_id)
+            data = response["data"]
+            url = response["url"]
+                    
+        return {"data": data, "url": url}
+
+
+    #
+    #  async_fetch_espn_all_leagues_data()
+    #    ESPN APIs returning all leagues quickly hit the API_LIMIT, so force use of tight date ranges
+    #      1. Get the team schedule from ESPN and determine next upcoming game
+    #      2. Call w/ date range up to upcoming game
+    #      2. Call w/ date range around upcoming game
+    #
+    async def async_fetch_espn_all_leagues_data(self, hass, lang) -> dict:
+        """Gets data from ESPN APIs for all leagues in specified sport."""
+
+        sensor_name = self.name
+        sport_path = self.sport_path
+        league_path = self.league_path
+        team_id = self.team_id.upper()
+
+        # Get date of next game
+        schedule_info = await self.async_get_team_schedule()
+        next_game_date = schedule_info.get("next_game_date") if schedule_info else None
+
+        # Narrow window: cover recent results and upcoming game if within 7 days
+        today_utc = datetime.now(timezone.utc).date()
+        day_before_yesterday = today_utc - timedelta(days=2)
+
+        d1 = day_before_yesterday.strftime("%Y%m%d")
+        if next_game_date and next_game_date <= today_utc + timedelta(days=7):
+            d2 = next_game_date.strftime("%Y%m%d")
+        else:
+            d2 = today_utc.strftime("%Y%m%d")
+
+        _LOGGER.debug(
+            "%s: All-league scoreboard call 1/1 dates=%s-%s (next_game=%s)",
+            sensor_name, d1, d2,
+            next_game_date.isoformat() if next_game_date else "unknown",
+        )
+
+        url_parms = {}
+        url_parms["lang"] = lang[:2]
+        url_parms["limit"] = str(API_LIMIT)
+        url_parms["dates"] = f"{d1}-{d2}"
+
+        url = URL_HEAD + sport_path + "/" + league_path + URL_TAIL
+
+        response = await async_call_espn_api(hass, url, url_parms, sensor_name, team_id)
+        data = response["data"]
+        url = response["url"]
+
+        # If event for team not returned, narrow date range and try again
+        if has_team(data, team_id) is False:
+            if (next_game_date and next_game_date > today_utc):
+                nd1 = (next_game_date - timedelta(days=1)).strftime("%Y%m%d")
+                nd2 = next_game_date.strftime("%Y%m%d")
+                if nd1 != d1 or nd2 != d2:  # avoid duplicate call
+                    _LOGGER.debug(
+                        "%s: All-league scoreboard call 2/2 dates=%s-%s (fallback to next game)",
+                        sensor_name, nd1, nd2,
+                    )
+
+                    url_parms["dates"] = f"{nd1}-{nd2}"
+                    url = URL_HEAD + sport_path + "/" + league_path + URL_TAIL
+
+                    response = await async_call_espn_api(hass, url, url_parms, sensor_name, team_id)
+                    data = response["data"]
+                    url = response["url"]
+
+        return {"data": data, "url": url}
+
+
+    #
+    #  async_update_values()
+    #
+    async def async_update_values(self, data) -> dict:
+        """Updates sensor values using data returned by API or in cache"""
+
+        sensor_name = self.name
+        league_id = self.league_id.upper()
+        team_id = self.team_id.upper()
+        lang = self.get_lang()
+
+        # Populate base values that do not need API data
+        values = {}
+        values = await async_clear_values()
+        if self.sport_path.lower() == "hockeytech":
+            values["sport"] = "hockey"
+        else:
+            values["sport"] = self.sport_path
+        values["sport_path"] = self.sport_path
+        values["league"] = league_id
+        values["league_path"] = self.league_path
+        values["league_logo"] = DEFAULT_LOGO
+        values["team_abbr"] = team_id
+        values["state"] = "NOT_FOUND"
+        values["last_update"] = arrow.now().format(arrow.FORMAT_W3C)
+        values["private_fast_refresh"] = False
+        values["api_url"] = self.api_url
+
+        # If there was an error (i.e. 404) w/ the API call...
+        if data is None:
+            values["api_message"] = "API error, no data returned"
+            _LOGGER.warning(
+                "%s: API did not return any data for team '%s'", sensor_name, team_id
+            )
+            return values
+
+        # When league_path is "all", parser needs league_map{} to do manual lookup
+        league_map = {}
+        if (self.league_path) == "all":
+            cache_key = f"{self.sport_path}:{self.league_path}:{self.team_id}"
+            team_cache = TeamTrackerCoordinator.all_team_cache.get(cache_key)
+            if team_cache:
+                league_map = team_cache.get("league_map", {})
+
+        values = await async_process_event(
+            values,
+            sensor_name,
+            data,
+            self.sport_path,
+            league_id,
+            DEFAULT_LOGO,
+            team_id,
+            league_map,
+            lang,
+        )
+
+        # If NOT_FOUND, try to get abbr w/ another API to make message easier to read
+        if (values["state"] == "NOT_FOUND" and 
+            is_integer(team_id)
+        ):
+            response = await self.provider.async_fetch_team_data(self.hass, self.sport_path, self.league_path)
+            teams = response["data"]
+            if teams:
+                team_abbr = next(
+                    (team["abbreviation"] for team in teams if team["id"] == team_id),
+                    None,
+                )
+            else:
+                team_abbr = None
+
+            values["team_id"] = team_id
+            if team_abbr:
+                values["team_abbr"] = team_abbr
+
+        return values
