+++ tests/conftest.py
@@ -102,10 +102,12 @@ async def mock_call_espn_api(hass):
     # Patch the actual utility function
     with patch("custom_components.teamtracker.utils.async_call_espn_api", new_callable=AsyncMock) as mock_utils, \
         patch("custom_components.teamtracker.config_flow.async_call_espn_api", new_callable=AsyncMock) as mock_cf, \
+        patch("custom_components.teamtracker.coordinator.async_call_espn_api", new_callable=AsyncMock) as mock_c, \
         patch("custom_components.teamtracker.espn.async_call_espn_api", new_callable=AsyncMock) as mock_espn, \
         patch("custom_components.teamtracker.async_call_espn_api", new_callable=AsyncMock) as mock:
         mock_utils.side_effect = _get_mock_api_data
         mock_cf.side_effect = _get_mock_api_data
+        mock_c.side_effect = _get_mock_api_data
         mock_espn.side_effect = _get_mock_api_data
         mock.side_effect = _get_mock_api_data
         yield mock
